Lab4/requirements.txt
Flask

Gunicorn

flask-sock

psycopg2

pyyaml